// src/interfaces/tokenInfo.ts
export interface ITokenInfo {
  symbol?: string;
  description?: string;
  explorer?: string;
  coinmarketcapUrl?: string;
  website_url?: string;
  name?: string;
  logo?: string;
  price?: string;
  price_date?: string;
  price_timestamp?: string;
  circulating_supply?: string;
  max_supply?: string;
  market_cap?: string;
  volume?: string;
  price_change?: string;
  price_change_pct?: string;
  volume_change?: string;
  volume_change_pct?: string;
  market_cap_change?: string;
  market_cap_change_pct?: string;
}